Refactor utilities to be a bit more organized.
