+++ shakenfist/blob.py
@@ -10,7 +10,8 @@ from shakenfist import constants
 from shakenfist import db
 from shakenfist import images
 from shakenfist import logutil
-from shakenfist import util
+from shakenfist.util import general as util_general
+from shakenfist.util import image as util_image
 
 
 LOG, _ = logutil.setup(__name__)
@@ -112,7 +113,7 @@ class Blob(dbo):
                 blob_path = os.path.join(
                     config.STORAGE_PATH, 'blobs', self.uuid)
 
-                info = images.identify(blob_path)
+                info = images.util_image.identify(blob_path)
                 info['mime-type'] = magic.Magic(mime=True).from_file(blob_path)
                 self._db_set_attribute('info', info)
 
@@ -129,8 +130,8 @@ def snapshot_disk(disk, blob_uuid, related_object=None):
     dest_path = os.path.join(config.STORAGE_PATH, 'blobs', blob_uuid)
 
     # Actually make the snapshot
-    with util.RecordedOperation('snapshot %s' % disk['device'], related_object):
-        images.snapshot(None, disk['path'], dest_path)
+    with util_general.RecordedOperation('snapshot %s' % disk['device'], related_object):
+        util_image.snapshot(None, disk['path'], dest_path)
         st = os.stat(dest_path)
 
     # And make the associated blob
