+++ shakenfist/daemons/external_api.py
@@ -3,7 +3,7 @@ import os
 from shakenfist.config import config
 from shakenfist.daemons import daemon
 from shakenfist import logutil
-from shakenfist import util
+from shakenfist.util import process as util_process
 
 
 LOG, _ = logutil.setup(__name__)
@@ -12,10 +12,10 @@ LOG, _ = logutil.setup(__name__)
 class Monitor(daemon.Daemon):
     def run(self):
         LOG.info('Starting')
-        util.execute(None, (config.API_COMMAND_LINE
-                            % {
-                                'port': config.API_PORT,
-                                'timeout': config.API_TIMEOUT,
-                                'name': daemon.process_name('api')
-                            }),
-                     env_variables=os.environ)
+        util_process.execute(None, (config.API_COMMAND_LINE
+                                    % {
+                                        'port': config.API_PORT,
+                                        'timeout': config.API_TIMEOUT,
+                                        'name': daemon.process_name('api')
+                                    }),
+                             env_variables=os.environ)
