+++ shakenfist/image_resolver/cirros.py
@@ -6,7 +6,7 @@ from shakenfist.config import config
 from shakenfist import exceptions
 from shakenfist.image_resolver import util as resolver_util
 from shakenfist import logutil
-from shakenfist import util
+from shakenfist.util import general as util_general
 
 
 LOG, _ = logutil.setup(__name__)
@@ -17,7 +17,7 @@ def resolve(name):
     if name == 'cirros':
         resp = requests.get(config.LISTING_URL_CIRROS,
                             allow_redirects=True,
-                            headers={'User-Agent': util.get_user_agent()})
+                            headers={'User-Agent': util_general.get_user_agent()})
         if resp.status_code != 200:
             raise exceptions.HTTPError(
                 'Failed to fetch %s, status code %d'
