+++ shakenfist/image_resolver/debian.py
@@ -6,7 +6,7 @@ from shakenfist.config import config
 from shakenfist import exceptions
 from shakenfist.image_resolver import util as resolver_util
 from shakenfist import logutil
-from shakenfist import util
+from shakenfist.util import general as util_general
 
 LOG, _ = logutil.setup(__name__)
 
@@ -23,7 +23,7 @@ def resolve(name):
         url = config.DOWNLOAD_URL_DEBIAN % {'vernum': vernum}
         checksum_url = config.CHECKSUM_URL_DEBIAN % {'vernum': vernum}
         resp = requests.head(url, allow_redirects=True,
-                             headers={'User-Agent': util.get_user_agent()})
+                             headers={'User-Agent': util_general.get_user_agent()})
         if resp.status_code != 200:
             raise exceptions.HTTPError(
                 'Failed to fetch %s, status code %d' % (url, resp.status_code))
@@ -33,7 +33,7 @@ def resolve(name):
         for vernum in range(9, 20):
             url = config.DOWNLOAD_URL_DEBIAN % {'vernum': vernum}
             resp = requests.head(url, allow_redirects=True,
-                                 headers={'User-Agent': util.get_user_agent()})
+                                 headers={'User-Agent': util_general.get_user_agent()})
             if resp.status_code != 200:
                 if found_any:
                     vernum -= 1
