+++ shakenfist/image_resolver/util.py
@@ -1,12 +1,12 @@
 import requests
 
 
-from shakenfist import util
+from shakenfist.util import general as util_general
 
 
 def fetch_remote_checksum(checksum_url):
     resp = requests.get(checksum_url,
-                        headers={'User-Agent': util.get_user_agent()})
+                        headers={'User-Agent': util_general.get_user_agent()})
     if resp.status_code != 200:
         return {}
