+++ shakenfist/tests/base.py
