+++ shakenfist/tests/test_tasks.py
@@ -2,10 +2,10 @@ import testtools
 
 from shakenfist import tasks
 from shakenfist import exceptions
-from shakenfist.tests import test_shakenfist
+from shakenfist.tests import base
 
 
-class TasksEqTestCase(test_shakenfist.ShakenFistTestCase):
+class TasksEqTestCase(base.ShakenFistTestCase):
     def test_QueueTask_eq(self):
         self.assertEqual(tasks.PreflightInstanceTask('abcd'),
                          tasks.PreflightInstanceTask('abcd'))
@@ -51,7 +51,7 @@ class TasksEqTestCase(test_shakenfist.ShakenFistTestCase):
                             tasks.FetchImageTask('http://someurm', 'fake_uuid'))
 
 
-class InstanceTasksTestCase(test_shakenfist.ShakenFistTestCase):
+class InstanceTasksTestCase(base.ShakenFistTestCase):
     def test_InstanceTask(self):
         i = tasks.InstanceTask('some-uuid')
         self.assertEqual('some-uuid', i.instance_uuid())
@@ -73,7 +73,7 @@ class InstanceTasksTestCase(test_shakenfist.ShakenFistTestCase):
             tasks.FetchImageTask(1234)
 
 
-class NetworkTasksTestCase(test_shakenfist.ShakenFistTestCase):
+class NetworkTasksTestCase(base.ShakenFistTestCase):
     def test_NetworkTask(self):
         n = tasks.NetworkTask('some-uuid')
         self.assertEqual('some-uuid', n.network_uuid())
